+++ bundles/digiroad/digiroad-personaldata/MyPlacesTab.js
@@ -93,7 +93,7 @@ function(instance, localization) {
     _showPlace : function(geometry, layerIdent) {
         // center map on selected place
         var center = geometry.getCentroid();
-        var mapmoveRequest = this.instance.sandbox.getRequestBuilder('MapMoveRequest')(center.x, center.y, geometry.getBounds(), false);
+        var mapmoveRequest = this.instance.sandbox.getRequestBuilder('MapMoveRequest')(center.x, center.y, geometry.getBounds());
         this.instance.sandbox.request(this.instance, mapmoveRequest);
 
         // add the myplaces layer to map