+++ bundles/liikennevirasto/lakapa-metadataflyout/instance.js
@@ -149,26 +149,7 @@ Oskari.clazz.define("Oskari.catalogue.bundle.metadataflyout.MetadataFlyoutBundle
      * @static
      *
      */
-    eventHandlers: {
-        'AfterMapLayerAddEvent': function (event) {
-            /* this might react when layer added */
-            /* this.scheduleShowMetadata(event.getMapLayer().getMetadataResourceUUID(); */
-        },
-        /**
-         * @method AfterMapLayerRemoveEvent
-         */
-        'AfterMapLayerRemoveEvent': function (event) {
-            /* this might react when layer removed */
-            /* this.scheduleShowMetadata(event.getMapLayer().getMetadataResourceUUID(); */
-        },
-        /**
-         * @method AfterMapLayerRemoveEvent
-         */
-        'AfterMapMoveEvent': function (event) {
-            /* this might react when map moved */
-
-        }
-    },
+    eventHandlers: {},
 
     /**
      * @method stop